# openapi-docs.yaml
openapi: 3.0.0
info:
  description: |
    This is the OpenAPI documentation for the TakeNote REST API. You can find
    out more at [https://app.swaggerhub.com/apis/vladpirlog/takenote-express/1.0.0]
    (https://app.swaggerhub.com/apis/vladpirlog/takenote-express/1.0.0).
  version: "1.0.0"
  title: TakeNote REST API
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: avi.pirlog@gmail.com
  license:
    name: MIT License
    url: "https://opensource.org/licenses/MIT"
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/vladpirlog/takenote-express/1.0.0
  - description: When running locally on port 8000
    url: "http://localhost:8000/"
...
paths:
...
  /auth/confirm:
    post:
      tags:
        - auth
        - tokens
      summary: Confirm email address
      operationId: confirm
      parameters:
        - name: token
          in: query
          description: The token to be used for confirmation
          required: true
          schema:
            $ref: "#/components/schemas/Token/properties/token"
      responses:
        "422":
          description: Input data missing from query
        "400":
          description: Couldn't confirm email address
        "200":
          description: Email address confirmed
        "202":
          description: A new confirmation token was sent to the user's email address
  /auth/request_confirmation:
    post:
      tags:
        - auth
        - tokens
      summary: Request a new confirmation token
      operationId: requestConfirmationToken
      security:
        - api_key: []
      responses:
        "401":
          description: No authenticated user
        "403":
          description: User has incompatible state
        "429":
          description: Email threshold has been reached
        "400":
          description: Couldn't issue new confirmation token
        "200":
          description: Confirmation token sent to the user's email address
...
  /notes:
    get:
      tags:
        - notes
      summary: User's notes
      operationId: allNotes
      security:
        - api_key: []
      parameters:
        - name: collaborations
          in: query
          required: false
          schema:
            type: boolean
            description: Specifies whether the notes where the user is collaborating should be included
      responses:
        "401":
          description: No authenticated user
        "403":
          description: User has incompatible state
        "422":
          description: Input data missing or invalid
        "200":
          description: Notes fetched
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - message
                  - notes
                properties:
                  status:
                    type: number
                    format: int32
                    description: HTTP status code
                  message:
                    type: string
                  notes:
                    type: array
                    items:
                      $ref: "#/components/schemas/Note"
                  collaboratingNotes:
                    type: array
                    items:
                      $ref: "#/components/schemas/Note"
    post:
      tags:
        - notes
      summary: Add note
      operationId: addNote
      security:
        - api_key: []
      responses:
        "401":
          description: No authenticated user
        "403":
          description: User has incompatible state
        "400":
          description: Couldn't create note
        "201":
          description: Note created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    format: int32
                    description: HTTP status code
                  message:
                    type: string
                  note:
                    $ref: "#/components/schemas/Note"
      requestBody:
        $ref: "#/components/requestBodies/AddOrUpdateNote"
  "/notes/{noteID}":
    get:
      tags:
        - notes
      summary: One user's note
      operationId: oneNote
      security:
        - api_key: []
      parameters:
        - name: noteID
          in: path
          description: ID of note to return
          required: true
          schema:
            $ref: "#/components/schemas/Note/properties/_id"
      responses:
        "401":
          description: No authenticated user
        "403":
          description: User has incompatible state
        "422":
          description: Input data missing from path
        "400":
          description: Couldn't get note
        "200":
          description: Note fetched
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    format: int32
                    description: HTTP status code
                  message:
                    type: string
                  note:
                    $ref: "#/components/schemas/Note"
    put:
      tags:
        - notes
      summary: Update note
      operationId: updateNote
      security:
        - api_key: []
      parameters:
        - name: noteID
          in: path
          description: ID of note to update
          required: true
          schema:
            $ref: "#/components/schemas/Note/properties/_id"
      responses:
        "401":
          description: No authenticated user
        "403":
          description: User has incompatible state
        "422":
          description: Input data missing from path
        "400":
          description: Couldn't update note
        "200":
          description: Note updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    format: int32
                    description: HTTP status code
                  message:
                    type: string
                  note:
                    $ref: "#/components/schemas/Note"
      requestBody:
        $ref: "#/components/requestBodies/AddOrUpdateNote"
    delete:
      tags:
        - notes
      summary: Delete note
      operationId: deleteNote
      security:
        - api_key: []
      parameters:
        - name: noteID
          in: path
          description: ID of note to delete
          required: true
          schema:
            $ref: "#/components/schemas/Note/properties/_id"
      responses:
        "401":
          description: No authenticated user
        "403":
          description: User has incompatible state
        "422":
          description: Input data missing from path
        "400":
          description: Couldn't delete note
        "200":
          description: Note deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    format: int32
                    description: HTTP status code
                  message:
                    type: string
...
  "/notes/{noteID}/share/collaborators":
    post:
      tags:
        - notes
        - collaborators
      summary: Add note collaborator
      operationId: addCollaborator
      security:
        - api_key: []
      parameters:
        - name: noteID
          in: path
          description: ID of note to add collab to
          required: true
          schema:
            $ref: "#/components/schemas/Note/properties/_id"
      responses:
        "401":
          description: No authenticated user
        "403":
          description: User has incompatible state
        "422":
          description: Input data missing from body or path
        "400":
          description: Couldn't add collaborator
        "200":
          description: Collaborator added
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    format: int32
                    description: HTTP status code
                  message:
                    type: string
                  permission:
                    $ref: "#/components/schemas/Permission"
      requestBody:
        $ref: "#/components/requestBodies/AddCollaborator"
  "/notes/{noteID}/share/collaborators/{permissionID}":
    delete:
      tags:
        - notes
        - collaborators
      summary: Delete note collaborator
      operationId: deleteCollaborator
      security:
        - api_key: []
      parameters:
        - name: noteID
          in: path
          description: ID of note to delete collab from
          required: true
          schema:
            $ref: "#/components/schemas/Note/properties/_id"
        - name: permissionID
          in: path
          description: ID of the permission to delete
          required: true
          schema:
            $ref: "#/components/schemas/Permission/properties/_id"
      responses:
        "401":
          description: No authenticated user
        "403":
          description: User has incompatible state
        "422":
          description: Input data missing from path
        "400":
          description: Couldn't delete collaborator
        "200":
          description: Collaborator deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    format: int32
                    description: HTTP status code
                  message:
                    type: string
...
components:
  schemas:
    Permission:
      type: object
      properties:
        _id:
          type: string
        subject:
          $ref: "#/components/schemas/User/properties/_id"
        level:
          type: number
          format: int32
          enum:
            - 0
            - 1
    Attachment:
      type: object
      properties:
        _id:
          type: string
        url:
          type: string
          description: A valid URL of a photo
        title:
          type: string
        description:
          type: string
    Tag:
      type: string
...
    Note:
      type: object
      required:
        - name
        - photoUrls
      properties:
        _id:
          type: string
        title:
          type: string
        content:
          type: string
        owner:
          $ref: "#/components/schemas/User/properties/_id"
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        shared:
          type: boolean
        shareCode:
          type: string
        archived:
          type: boolean
        color:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      xml:
        name: Note
  requestBodies:
...
    AddOrUpdateNote:
      content:
        multipart/formdata:
          schema:
            type: object
            properties:
              title:
                $ref: "#/components/schemas/Note/properties/title"
              content:
                $ref: "#/components/schemas/Note/properties/content"
              archived:
                $ref: "#/components/schemas/Note/properties/archived"
              color:
                $ref: "#/components/schemas/Note/properties/color"
    AddCollaborator:
      content:
        multipart/formdata:
          schema:
            type: object
            required:
              - user
              - type
            properties:
              user:
                type: string
                description: Username, email or id of the user to make collab
              type:
                type: string
                enum:
                  - r
                  - rw
...
  securitySchemes:
    api_key:
      type: apiKey
      description: JWT signed token used for authentication
      name: api_key
      in: cookie